+++ CHANGELOG.md
@@ -9,6 +9,11 @@ All notable changes to MiniJinja are documented here.
 - Fixed an incorrect error case in `call_method`.  Now `UnknownMethod`
   is returned instead of `InvalidOperation` correctly. #439
 
+- Added `Environment::set_unknown_method_callback` which allows a user
+  to intercept method calls on primitives.  The motivation here is that
+  this can be used to implement python like methods to improve the
+  compatibility with Jinja2 Python templates.  #441
+
 ## 1.0.14
 
 - Fixed a bug with broken closure handling when working with nested